+++ apps/frontpage/components/home/manager/component-image.tsx
