+++ apps/frontpage/components/home/manager/slide-controls.tsx
