+++ apps/frontpage/components/home/manager/slide-docs.tsx
