+++ apps/frontpage/components/home/manager/slide-interactions.tsx
