+++ apps/frontpage/components/home/manager/tabs.tsx
